+++ lib/vow.ex
@@ -3,6 +3,7 @@ defmodule Vow do
   TODO
   """
 
+  import Kernel, except: [get_in: 2, update_in: 3, put_in: 3]
   alias Vow.{Conformable, ConformError}
 
   @typedoc """
@@ -50,6 +51,70 @@ defmodule Vow do
   end
 
   defdelegate unform(vow, value), to: Vow.Conformable
+
+  @doc """
+  """
+  @spec get_in(t, path :: [term]) :: t | nil
+  def get_in(vow, []), do: vow
+  def get_in(vow, path) do
+    try do
+      Kernel.get_in(vow, Enum.map(path, &lazy_path/1))
+    rescue
+      _ -> nil
+    end
+  end
+
+  @doc """
+  """
+  @spec update_in(t, path :: [term], (term -> term)) :: t
+  def update_in(vow, [], fun), do: fun.(vow)
+  def update_in(vow, path, fun) do
+    try do
+      Kernel.update_in(vow, Enum.map(path, &lazy_path/1), fun)
+    rescue
+      _ -> vow
+    end
+  end
+
+  @doc """
+  """
+  @spec put_in(t, path :: [term], value :: term) :: t
+  def put_in(vow, path, value) do
+    update_in(vow, path, fn _ -> value end)
+  end
+
+  @doc false
+  @spec lazy_path(term) :: Access.access_fun(t, term)
+  defp lazy_path(fun) when is_function(fun, 3), do: fun
+  defp lazy_path(i) when is_integer(i) do
+    fn
+      type, data, next when is_list(data) ->
+        Access.at(i).(type, data, next)
+      type, data, next when is_tuple(data) ->
+        Access.elem(i).(type, data, next)
+      type, data, next ->
+        Access.key(i).(type, data, next)
+    end
+  end
+  defp lazy_path(key) do
+    fn type, data, next ->
+      Access.key(key).(type, data, next)
+    end
+  end
+
+  @type generator :: Vow.Generatable.generator
+  @type gen_fun :: (() -> generator)
+  @type override :: {path :: [term], gen_fun}
+
+  @doc """
+  """
+  @spec gen(t, [override]) :: {:ok, generator} | {:error, reason :: term}
+  def gen(vow, overrides \\ []) do
+    Enum.reduce(overrides, vow, fn {path, gen_fun}, acc ->
+      put_in(acc, path, gen_fun.())
+    end)
+    |> Vow.Generatable.gen()
+  end
 
   @doc """
   """