+++ lib/vow/maybe.ex
@@ -1,5 +1,6 @@
 defmodule Vow.Maybe do
   @moduledoc false
+  @behaviour Access
 
   defstruct vow: nil
 
@@ -10,6 +11,21 @@ defmodule Vow.Maybe do
   @spec new(Vow.t()) :: t
   def new(vow) do
     %__MODULE__{vow: vow}
+  end
+
+  @impl Access
+  def fetch(%__MODULE__{vow: vow}, key) do
+    Access.fetch(vow, key)
+  end
+
+  @impl Access
+  def get_and_update(%__MODULE__{vow: vow}, key, fun) do
+    Access.get_and_update(vow, key, fun)
+  end
+
+  @impl Access
+  def pop(%__MODULE__{vow: vow}, key) do
+    Access.pop(vow, key)
   end
 
   defimpl Vow.RegexOperator do
@@ -70,6 +86,30 @@ defmodule Vow.Maybe do
 
     def unform(vow, value) do
       {:error, %Vow.UnformError{vow: vow, value: value}}
+    end
+  end
+
+  if Code.ensure_loaded?(StreamData) do
+    defimpl Vow.Generatable do
+      @moduledoc false
+
+      @impl Vow.Generatable
+      def gen(vow) do
+        case @protocol.gen(vow.vow) do
+          {:error, reason} -> {:error, reason}
+          {:ok, data} ->
+            if Vow.regex?(vow.vow) do
+              {:ok, StreamData.map(
+                StreamData.list_of(data, length: 0..1),
+                fn
+                  [x] when is_list(x) -> x
+                  otherwise -> otherwise
+                end)}
+            else
+              {:ok, StreamData.list_of(data, length: 0..1)}
+            end
+        end
+      end
     end
   end
 end