+++ lib/vow/nilable.ex
@@ -1,5 +1,6 @@
 defmodule Vow.Nilable do
   @moduledoc false
+  @behaviour Access
 
   defstruct [:vow]
 
@@ -10,6 +11,21 @@ defmodule Vow.Nilable do
   @spec new(Vow.t()) :: t
   def new(vow) do
     %__MODULE__{vow: vow}
+  end
+
+  @impl Access
+  def fetch(%__MODULE__{vow: vow}, key) do
+    Access.fetch(vow, key)
+  end
+
+  @impl Access
+  def get_and_update(%__MODULE__{vow: vow}, key, fun) do
+    Access.get_and_update(vow, key, fun)
+  end
+
+  @impl Access
+  def pop(%__MODULE__{vow: vow}, key) do
+    Access.pop(vow, key)
   end
 
   defimpl Vow.Conformable do
@@ -27,5 +43,23 @@ defmodule Vow.Nilable do
     @impl Vow.Conformable
     def unform(_vow, nil), do: {:ok, nil}
     def unform(%@for{vow: vow}, value), do: @protocol.unform(vow, value)
+  end
+
+  if Code.ensure_loaded?(StreamData) do
+    defimpl Vow.Generatable do
+      @moduledoc false
+
+      @impl Vow.Generatable
+      def gen(vow) do
+        case @protocol.gen(vow.vow) do
+          {:error, reason} -> {:error, reason}
+          {:ok, data} ->
+            {:ok, StreamData.one_of([
+              StreamData.constant(nil),
+              data
+            ])}
+        end
+      end
+    end
   end
 end